Add getRoadPathFromPos + move bunker functions
Bunker functions moved to data to prevent circular dependency issues on
roomDesign
